requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
django-credit-cards==0.4.1
django-filter==24.2
django-jazzmin==3.0.0
django-unfold==0.30.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.6
pillow==10.3.0
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.12.2
